gen_data/requirements.txt
# Pillow==6.2.1
Pillow==10.0.1